Split Equilibrium.py into per-class modules
Break the 692-line monolithic src/Equilibrium.py into six focused
modules named after their sole class (lowercase per PEP8):

  src/coils.py         -> Coils
  src/equilibrium.py   -> Equilibrium (depends on coils)
  src/mesh.py          -> Mesh
  src/plotting.py      -> PlotEquilibrium
  src/resonances.py    -> Resonances
  src/sources.py       -> Sources

Class bodies are unchanged; each module carries only the imports it
actually needs. Static dependency is one-way: equilibrium -> coils.
Mesh, PlotEquilibrium, Resonances, and Sources receive their peers as
constructor arguments at runtime, so no cross-imports are required.

The matplotlib.rcParams.update(...) that lived at the top of
Equilibrium.py is now in plotting.py. main.py imports plotting, so
its call chain still applies the styling. Modules that use matplotlib
internally (mesh, resonances) will render with default rcParams if
imported without plotting; acceptable given no other caller currently
does that.

Updated imports:
  main.py:          from equilibrium import Equilibrium
                    from mesh import Mesh
                    from plotting import PlotEquilibrium
  ChargedMotion.py: from equilibrium import Equilibrium

tests/test_smoke.py now parametrizes over each of the six new modules
and asserts each exposes its class.

Verified locally: ruff check + format --check clean; pytest 8/8 pass.
